Modification pour diminuer le temps d'execution et éviter d'avoir le clic droit envoyé par le systeme
https://msdn.microsoft.com/en-us/library/windows/desktop/ms644985(v=vs.85).aspx
https://msdn.microsoft.com/en-us/library/windows/desktop/ms644986(v=vs.85).aspx

he hook procedure should process a message in less time than the data entry specified in the LowLevelHooksTimeout value in the following registry key:
HKEY_CURRENT_USER\Control Panel\Desktop

The value is in milliseconds. If the hook procedure times out, the system passes the message to the next hook. However, on Windows 7 and later, the hook is silently removed without being called. There is no way for the application to know whether the hook is removed.